requirements.txt
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
blinker==1.9.0
blis==1.3.3
cachetools==6.2.6
catalogue==2.0.10
celery==5.6.2
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
distro==1.9.0
Django==6.0.2
fastapi==0.128.5
gitdb==4.0.12
GitPython==3.1.46
groq==1.0.0
google-generativeai==0.8.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
MarkupSafe==3.0.3
murmurhash==1.0.15
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pdfminer3==2018.12.3.0
pillow==12.1.0
plotly==6.5.2
preshed==3.0.12
prompt_toolkit==3.0.52
pyarrow==23.0.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==81.0.0
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
starlette==0.52.1
streamlit==1.54.0
tenacity==9.1.4
thinc==8.3.10
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
whitenoise==6.11.0
gunicorn==23.0.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
wrapt==2.1.1
